Extract `shift` to Dhall.Shift
Context: #2313
