+++ dhall/src/Dhall/Normalize.hs
@@ -14,7 +14,7 @@ module Dhall.Normalize (
     , ReifiedNormalizer (..)
     , judgmentallyEqual
     , subst
-    , Syntax.shift
+    , shift
     , isNormalized
     , isNormalizedWith
     , freeIn
@@ -28,6 +28,7 @@ import Data.Sequence         (ViewL (..), ViewR (..))
 import Data.Traversable
 import Instances.TH.Lift     ()
 import Prelude               hiding (succ)
+import Dhall.Shift           (shift)
 
 import Dhall.Syntax
     ( Binding (Binding)
@@ -69,18 +70,18 @@ subst (V x n) e (Lam cs (FunctionBinding src0 y src1 src2 _A) b) =
     Lam cs (FunctionBinding src0 y src1 src2 _A') b'
   where
     _A' = subst (V x n )                         e  _A
-    b'  = subst (V x n') (Syntax.shift 1 (V y 0) e)  b
+    b'  = subst (V x n') (shift 1 (V y 0) e)  b
     n'  = if x == y then n + 1 else n
 subst (V x n) e (Pi cs y _A _B) = Pi cs y _A' _B'
   where
     _A' = subst (V x n )                         e  _A
-    _B' = subst (V x n') (Syntax.shift 1 (V y 0) e) _B
+    _B' = subst (V x n') (shift 1 (V y 0) e) _B
     n'  = if x == y then n + 1 else n
 subst v e (Var v') = if v == v' then e else Var v'
 subst (V x n) e (Let (Binding src0 f src1 mt src2 r) b) =
     Let (Binding src0 f src1 mt' src2 r') b'
   where
-    b' = subst (V x n') (Syntax.shift 1 (V f 0) e) b
+    b' = subst (V x n') (shift 1 (V f 0) e) b
       where
         n' = if x == f then n + 1 else n
 
@@ -194,9 +195,9 @@ normalizeWithM ctx e0 = loop (Syntax.denote e0)
               case f' of
                 Lam _ (FunctionBinding _ x _ _ _A) b₀ -> do
 
-                    let a₂ = Syntax.shift 1 (V x 0) a'
+                    let a₂ = shift 1 (V x 0) a'
                     let b₁ = subst (V x 0) a₂ b₀
-                    let b₂ = Syntax.shift (-1) (V x 0) b₁
+                    let b₂ = shift (-1) (V x 0) b₁
 
                     loop b₂
                 _ ->
@@ -256,7 +257,7 @@ normalizeWithM ctx e0 = loop (Syntax.denote e0)
                         pure (TextLit (Chunks [] (Text.pack (show n))))
                     App (App ListBuild _A₀) g -> loop (App (App (App g list) cons) nil)
                       where
-                        _A₁ = Syntax.shift 1 "a" _A₀
+                        _A₁ = shift 1 "a" _A₀
 
                         list = App List _A₀
 
@@ -383,9 +384,9 @@ normalizeWithM ctx e0 = loop (Syntax.denote e0)
                             Just app' -> loop app'
     Let (Binding _ f _ _ _ r) b -> loop b''
       where
-        r'  = Syntax.shift   1  (V f 0) r
+        r'  = shift   1  (V f 0) r
         b'  = subst (V f 0) r' b
-        b'' = Syntax.shift (-1) (V f 0) b'
+        b'' = shift (-1) (V f 0) b'
     Annot x _ -> loop x
     Bool -> pure Bool
     BoolLit b -> pure (BoolLit b)
